+++ .gitignore
@@ -1,0 +1,73 @@
+# Byte-compiled / optimized / DLL files
+__pycache__/
+*.py[cod]
+*.pyo
+*.egg-info/
+*.egg
+
+# C extensions
+*.so
+
+# Distribution / packaging
+.Python
+.venv/
+venv/
+.vscode/
+build/
+develop-eggs/
+dist/
+downloads/
+eggs/
+.eggs/
+lib/
+lib64/
+parts/
+sdist/
+var/
+*.egg-info/
+.installed.cfg
+*.egg
+
+# PyInstaller
+#  Usually these files are written by a python script from a template
+#  before PyInstaller builds the exe, so as to inject date/other infos into it.
+*.manifest
+*.spec
+
+# Installer logs
+pip-log.txt
+pip-delete-this-directory.txt
+
+# Unit test / coverage reports
+htmlcov/
+.tox/
+.coverage
+.coverage.*
+.cache
+nosetests.xml
+coverage.xml
+*.cover
+.hypothesis/
+.pytest_cache/
+
+# Jupyter Notebook
+.ipynb_checkpoints
+
+# IDEs and editors
+# VSCode
+.vscode/
+# PyCharm
+.idea/
+
+# Operating systems
+*.swp
+*.swo
+.DS_Store
+._*
+.Spotlight-V100
+.Trashes
+
+# Custom exclusions
+# You can add custom exclusions specific to your project here
+# e.g., exclude a secrets file
+# secrets.yml
