+++ models.py
@@ -1,0 +1,5 @@
+class Student:
+    def __init__(self, roll, name, percentage):
+        self.roll = roll
+        self.name = name
+        self.percentage = percentage
